// src/resources/status/subscribers.ts
// File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.

import { APIResource } from '../../core/resource';
import { APIPromise } from '../../core/api-promise';
import { buildHeaders } from '../../internal/headers';
import { RequestOptions } from '../../internal/request-options';

export class Subscribers extends APIResource {
  /**
   * Get subscriber count
   */
  getCount(options?: RequestOptions): APIPromise<void> {
    return this._client.get('/status/subscribers/count', {
      ...options,
      headers: buildHeaders([{ Accept: '*/*' }, options?.headers]),
    });
  }
}